rename hf_lit_module.py to hf_auto_model.py
