+++ nemo/collections/llm/gpt/model/__init__.py
@@ -37,7 +37,7 @@ from nemo.collections.llm.gpt.model.gemma import (
     GemmaConfig7B,
     GemmaModel,
 )
-from nemo.collections.llm.gpt.model.hf_lit_module import HfAutoModel
+from nemo.collections.llm.gpt.model.hf_auto_model import HfAutoModel
 from nemo.collections.llm.gpt.model.llama import (
     CodeLlamaConfig7B,
     CodeLlamaConfig13B,
