+++ nemo/collections/llm/recipes/hf_auto_model.py
@@ -24,7 +24,7 @@ from nemo import lightning as nl
 from nemo.collections import llm
 from nemo.collections.llm.api import finetune, pretrain
 from nemo.collections.llm.gpt.data.mock import MockDataModule
-from nemo.collections.llm.gpt.model.hf_lit_module import HfAutoModel
+from nemo.collections.llm.gpt.model.hf_auto_model import HfAutoModel
 from nemo.collections.llm.recipes.log.default import default_log, default_resume, tensorboard_logger
 from nemo.collections.llm.recipes.optim.adam import pytorch_adam_with_cosine_annealing
 from nemo.utils.exp_manager import TimingCallback
